Regenerate stubs from source with stubgen
The previous stubs had been auto-generated but were throwing up a lot
of errors from stubtest. I haven't traced the exact details of the
error, but I suspect a version mismatch. This set of stubs has been
generated with the version of the library that is tested against in
CI, so should remain consistent going forward.